Fixed Last Name Omission
Made an enhancement in which Launchpad automatically detects if a customer has provided their last name.
If so, it proceeds as normal with customer entry.
If not, it enters the customer's name as the company name instead.

This prevents Evosus from compulsively looking up thousands of database matches, saving on profile entry time.